Create reusable fieldrow component
