+++ src/pages/projectSettings/tabs/VariablesTab/VariableRow.tsx
@@ -2,10 +2,9 @@ import styled from "@emotion/styled";
 import { palette } from "@leafygreen-ui/palette";
 import { ObjectFieldTemplateProps } from "@rjsf/core";
 import Icon from "components/Icon";
+import { getFields } from "components/SpruceForm/utils";
 import { size } from "constants/tokens";
-import { form } from "../utils";
 
-const { getFields } = form;
 const { yellow } = palette;
 
 export const VariableRow: React.VFC<
